Migrates the Docker container from pip to Conda with the exception of a few packages that are not available or up-to-date in the Conda repositories.
Also updates some Python packages and cleans-up the Dockerfile some.

Additionally, the nbconvert config file is updated to disable xsrf token
checking in Jupyter so that tests will pass. One of the updated libraries
seems to have made the tokens a requirement by default. The tokens are
sometimes present in the request headers when running Datalab locally,
but the tokens are not being sent in tests for whatever reason. This
will be tracked via an issue and handled in a follow-up.